Add division symbol to the display when clicked
The reason why operator's value is / but ÷ is displayed is because
the computer will likely need to read operator's value. If the computer
reads ÷ it will probably throw an error because its division operator
is /. However, I think ÷ is easier for the average person to
understand. So the computer sees / but the user sees ÷.