Add the condition that num-two can't be nan
A bug would allow operate to be called even if numTwo was not given
a value. After watching variables it was realized that numTwo kept
equating to NaN when it was not given a value.

This bug was fixed by only allowing operate to be called if numTwo
did not equal undefined or NaN.