Add the class number to all number buttons
Allows for the creation of a nodelist containing all of the number
buttons, which then allows for iteration through the nodelist and
applying the same functionality to all number buttons. This is much
more convenient than having to create a reference to every number
button and giving them the same code to run.