Add the class operator to some of the op buttons
This will allow for the creation of a nodelist and iterating through
that nodelist to apply the same functions to all of the operators.
Only the operators for addition, subtraction, multiplication, and
division were given this class because they will certainly have almost
the exact same functionality, if not the exact same.