Initialize additional-numbers variable
Any numbers typed in after the operation is complete should not be
appended to calcPara, but they do have to be saved so that numTwo can
be given their value. All attempted methods other than pushing the
additional numbers into an array only returned one number, so this is
the method that will be used. To do that, an array is necessary.