Check if operator is true and other cond is false
Checking if operator is true and doesNumTwoContainDecimal is false, and
appending a decimal point to calcPara's textContent if these conditions
are met allows the user to input a decimal point anywhere inside of
numTwo. Setting doesNumTwoContainDecimal to true when the conditions
are met prevents the user from spamming decimal points inside of
numTwo.