Initialize function reset-all-info
Realized that because there are quite a few instances where variables
are reset manually, it would be best to create a function that does
this instead. It is easier to remember to call a function than is is to
remember every single variable that has to be reset and the values
they should be set to. It will also make the code more readable.