Check if additional-numbers contains a decimal
Check if additionalNumbers contains a decimal point and prevent the if
statement from pushing in a decimal point if it does. Fixes a bug that
would empty out calcPara's textContent or convert numTwo into an
integer even if it was a decimal, if additionalNumbers had more than
one decimal point in it.