Change op-input to division symbol
If e.key equals forward slash change the value of opInput to the
division symbol. The division symbol does not exist on the average
keyboard, but it is more familiar to english speakers. This will be
used to allow the user to input forward slash and output the division
symbol, which they are likely more familiar with.