Change op-input to multiplication symbol
Else if e.key equals asterisk set opInput to asterisk. Much like the
division symbol, the multiplication symbol does not exist on the
average keyboard. However the user is likely more familiar with it
than using an asterisk for multiplication. This will be used to allow
the user to input an asterisk and output the multiplication symbol.