Initialize function operators-functionality
Initialize function operatorsFunctionality with parameter input. Since
the functionality for operators can be reused for at least some parts
of adding keyboard support to the operators, it would be more
convenient to create a function and reuse it when needed. It has been
given the parameter input because the code that is going to be pasted
inside of this function requires input for many things.