Check that final char in calc-para is a decimal
Add a new condition to else if statement inside of backBtn's event
listener that checks if final character in calcPara is a decimal.
Need to know if its a decimal so that if it is, the variable
doesNumOneContainDecimal or doesNumTwoContainDecimal can be reset to
false and fix a bug that prevents the user from adding decimals if the
decimal of the current number was previously erased using backBtn.