Slice from 0 to negative 3 if operator
When backBtn is clicked and the last character in calcPara textContent
is an operator, slice from 0 to -3. Fixes a very agitating bug that
makes numTwo return as NaN if the operator was previously erased using
backBtn, because an extra space threw the index off and would make
numTwo try to use the operator.