Add prevent default to all buttons
Fixes an agitating issue that would cause the buttons to enter
keyboard focus mode when using a mix of keyboard and onscreen input.
This caused many issues. For example, it would clear the user's input
instead of calling operate when they press enter because clear or back
were keyboard focused.