Reorganize code and replace if statement condition
Reorganize code and replace if statement condition with a call to
checkForDividingByZero. Replace the code inside of the if statement
with a call to noDividingByZero and move the code that was there
previously into an else statement.

Fixes a bug that does not prevent the user from dividing by zero if
hasOperateBeenCalled is true.