Reset additional-numbers inside of reset-all-info
If the user adds to additionalNumbers before clearing display by
clicking backBtn, operate will be called when the user attempts to
add an operator to the equation. This will display an equation that
seemingly has no numTwo, but it actually does because additionalNumbers
is pushed into numTwo despite the display being cleared. Resetting
additionalNumbers when resetAllInfo is called fixes this bug.